requirements.txt
absl-py==2.3.1
albumentations==1.4.6
attrs==25.3.0
black==24.10.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
email-validator==2.2.0
eventlet==0.34.3
fastapi==0.115.0
filelock==3.19.1
Flask==3.1.2
flask-login==0.6.3
flask-socketio==5.3.5
flask-wtf==1.2.2
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.9.0
git-filter-repo==2.47.0
google-generativeai
gradio==4.44.0
streamlit==1.40.0
gunicorn==21.2.0
h5py==3.12.1
hf-xet==1.1.10
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.3
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
pillow==10.2.0
protobuf==4.25.8
pycparser==2.23
pyparsing==3.2.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pyttsx3==2.99
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
reportlab==4.0.7
requests==2.32.5
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
seaborn==0.13.2
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.32.1
wandb==0.18.7
Werkzeug==3.1.3
xgboost==3.0.5
yt-dlp==2024.8.6
